feat: make sentry as opt in - after feedback changes
