+++ apps/trading/components/welcome-dialog/welcome-notice-dialog.tsx
@@ -10,7 +10,6 @@ import {
   useLinks,
 } from '@vegaprotocol/environment';
 import { ProposedMarkets } from './proposed-markets';
-import { TelemetryApproval } from './telemetry-approval';
 
 export const WelcomeNoticeDialog = () => {
   const { VEGA_ENV } = useEnvironment();
@@ -56,9 +55,6 @@ export const WelcomeNoticeDialog = () => {
           <ExternalLink target="_blank" href={BLOG}>
             {t('Read about the mainnet launch')}
           </ExternalLink>
-        </li>
-        <li>
-          <TelemetryApproval />
         </li>
       </ul>
       <ProposedMarkets />
